Found by https://discourse.julialang.org/t/how-to-optimize-computation-within-vectorized-list-operation-and-large-array/89008
```
julia> xc64 = rand(ComplexF64, 1024); yc64=similar(xc64);
#Old
julia> @benchmark @. $yc64 = inv($xc64)
BenchmarkTools.Trial: 10000 samples with 6 evaluations.
 Range (min … max):  5.467 μs …  10.603 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     5.983 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.838 μs ± 368.716 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █▂                   ██▁                                    ▂
  ██▄▅▄▄▃▁▅▆▇▇▆▇▆▆▆▇▇▇▇███▇▇▇█▇███▇▆▆▆▅▆▅▇▆▆▇▇▅▄▅▃▅▄▅▃▅▄▁▃▃▁▃ █
  5.47 μs      Histogram: log(frequency) by time      6.86 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
#new
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  1.937 μs …   5.240 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     2.041 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.052 μs ± 133.940 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

          ▂█▆▁    ▅▃█▁                                         
  ▁▂▅▆▄▂▁▂████▆▃▄█████▃▃▂▂▂▇▅▃▃▂▃▂▂▁▁▁▁▁▁▁▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▃
  1.94 μs         Histogram: frequency by time        2.32 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> xc64 = 1e200.+rand(ComplexF64, 1024); yc64=similar(xc64);
#old
julia> @benchmark @. $yc64 = inv($xc64)
@b^[[ABenchmarkTools.Trial: 10000 samples with 6 evaluations.
 Range (min … max):  5.465 μs …  10.222 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     5.482 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.749 μs ± 360.455 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █                          ▇▂                               ▁
  █▆▃▄▄▅▃▄▄▅▅▅▆▇▅▆▆▆▆▇▆▇▄▅▆▆▆██▇▆▇▇▇▇█▆▇▇▇▆▆▇▇▇▆▆▆▅▆▆▆▆▆▄▅▅▄▄ █
  5.46 μs      Histogram: log(frequency) by time      6.57 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
#new
julia> @benchmark @. $yc64 = inv($xc64)
BenchmarkTools.Trial: 10000 samples with 6 evaluations.
 Range (min … max):  5.697 μs …  14.716 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     5.727 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.850 μs ± 454.134 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▇█                     ▅▃                                   ▁
  ███▅▄▁▄▃▃▅▅▆▇▇▇▆▇▆▇▇▇▇███▇▇▇▇▇▇▇▇▆▆▆▆▆▅▆▆▆▄▅▆▆▆▅▆▆▅▄▅▅▅▃▄▅▆ █
  5.7 μs       Histogram: log(frequency) by time      7.09 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
So for most values that will appear in practice this is 2x faster (but it's 20% slower for values that risk overflow.