@mikmoore 2022-10-20
Orthogonal to the changes you've made (but among the same code), I see a benefit to the following substitution
```julia
function robust_cinv(c::Float64, d::Float64)
    r = d/c
    z = muladd(d, r, c)
    p = one(r)/z
    q = -r/z
    return p, q
end
```
The innovation here is to permit the `p` and `q` divisions to be done with a vectorized division rather than a sequential divide and multiply. This seems to perform slightly better and might avoid some minor inaccuracies compared to the existing invert-and-multiply for `q`.

Without your change, this improves the benchmarked runtime from 7.6ns to 5.4ns on my machine. For comparison, your version on your new fastpath benchmarks at 4.1ns.