@oscardssmith 2022-10-20
Vectorized divide should almost always be slower.