@mikmoore 2022-10-20
Of course division is slower than multiplication. But what's at question here is doing two divisions simultaneously versus doing a scalar division *followed by* a scalar multiplication that depends on the result.

I benchmarked the change and it was definitely faster for me. Did you see something different? Using your benchmark on my machine, now:
```julia
julia> xc64 = rand(ComplexF64, 1024); yc64=similar(xc64);

# v1.8.0
julia> @benchmark @. $yc64 = inv($xc64)
BenchmarkTools.Trial: 10000 samples with 5 evaluations.
 Range (min … max):  6.640 μs … 22.440 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     6.720 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   6.835 μs ±  1.002 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █▄▂▁                                                       ▁
  ████▆▆▆▄▄▄▃▄▁▃▁▁▁▁▃▃▁▄▁▃▁▁▃▁▅▅▆▅▃▄▄▄▁▄▁▃▃▅▁▃▄▃▃▁▁▁▁▃▁▁▁▁▄▅ █
  6.64 μs      Histogram: log(frequency) by time     12.4 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.

# changes to robust_cinv
julia> @benchmark @. $yc64 = inv($xc64)
BenchmarkTools.Trial: 10000 samples with 6 evaluations.
 Range (min … max):  6.050 μs …  23.033 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     6.183 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   6.340 μs ± 977.819 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   █
  ▆█▆▅▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▁▁▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂▁▂▁▂▂▂▁▂▂▂▂▂▂▂▁▂▂▁▁▂ ▂
  6.05 μs         Histogram: frequency by time        11.8 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.

# this PR
julia> @benchmark @. $yc64 = inv($xc64)
BenchmarkTools.Trial: 10000 samples with 7 evaluations.
 Range (min … max):  4.971 μs …  15.957 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     5.014 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.126 μs ± 705.331 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █▃▇                                                         ▁
  ████▆█▅▄▄▃▃▁▁▁▁▁▃▁▄▁▃▁▃▁▃▃▁▁▄▆▆▅▄▄▃▄▃▁▃▃▄▄▅▁▃▃▃▄▁▃▃▄▁▃▃▁▁▄▄ █
  4.97 μs      Histogram: log(frequency) by time       9.1 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
```