@mikmoore 2022-11-07
One philosophical nit:
Without loss of generality, assume `abs(d) <= abs(c)` and `0 < abs(c)` and assume `fma` is not used by the `muladd` calls (although this argument only depends on it being applied consistently). My tiny anxiety with this fast path is that, depending purely on the scale of the argument, `inv(complex(c,d))` is computed as either `complex(c,-d) / (d^2 + c^2)` (new fast path) or as `complex(1, -d/c) / (d/c*d + c)` (slow path, ignoring the anti-overflow scaling -- but the scales are powers of 2 so would not change the result). This introduces the possibility that `inv(x)/s != inv(x*s)` even when this would have been a preventable disagreement (eg when `s = 2.0^n`). But by the time one is into complex arithmetic, I don't think one necessarily expects this to hold. It can introduce some irritating edge cases, although I can't really assert they're anything someone would care about.

That said, the current slow path is frustratingly slow -- especially since it's unnecessary for "typical" numbers. Further, I would imagine that the fast path version is generally more accurate. I think this PR is ultimately an improvement.

Unless someone can raise a credible and practical reason to share my anxiety on the differing computational paths, this looks good to me.