@oscardssmith 2022-11-08
Yeah I really wish I could think of a better slow path which would let us delete the fast path, but I somewhat doubt it exists.